# dsp-integration/docs/dsp-integration/api/openapi.yaml
openapi: 3.1.0
info:
  title: Personalisation Hub — Display Types & DSP Integration API
  version: 1.0.0-draft
  description: |
    The agreed contract for every API added by the Display Types & DSP
    Integration build. Implement exactly these endpoints, fields and error
    codes; do not add endpoints, fields or behaviour that are not here.
    Changes to this file are made in the same commit as the code change.

    Two surfaces:
      * Partner API (/v1)       — DSPs and direct advertisers (tier 2).
      * Admin API (/admin/v1)   — behind HQ Admin screens.
    Plus the public sellers.json. The OpenRTB bid request/response contract
    is documented in API.md (it is sent by us, not served).

    Out of scope (existing Personalisation Hub functionality, no endpoints
    here): playback, playlists' playback behaviour, targeting evaluation,
    distribution to players, playback logging, campaign playback analytics.

    Jobs with no endpoint: the SSP auction runs as a scheduled job
    (CLI for demos: npm run auction:run); billing line items are stored
    only (CLI to print them: npm run billing:print). Neither has a UI,
    report or API.
servers:
  - url: https://{instance}/api
    variables:
      instance:
        default: hub.example-retailer.com
        description: The retailer's own Personalisation Hub instance (in its VPC).
...
security:
  - partnerToken: []
paths:
...
  /v1/campaigns:
    post:
      tags: [Partner — Campaigns]
      operationId: createCampaign
      summary: Create a campaign (content package) — a mandatory default layer plus optional targeted versions
      description: |
        Targeting rules are validated against the variables permitted for the
        calling DSP and stored in the existing campaign targeting structure.
        Evaluation and playback are the existing platform's.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CampaignCreate'}
      responses:
        '201': {description: Created, content: {application/json: {schema: {$ref: '#/components/schemas/CampaignStatus'}}}}
        '400': {$ref: '#/components/responses/ValidationFailed'}
        '401': {$ref: '#/components/responses/Unauthorised'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '422': {$ref: '#/components/responses/VariableNotPermitted'}
  /v1/campaigns/{campaignId}/assets:
    post:
      tags: [Partner — Campaigns]
      operationId: uploadAsset
      summary: Upload creative for the default layer or a targeted version; runs automated checks
      parameters: [{$ref: '#/components/parameters/CampaignId'}]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [version, file]
              properties:
                version: {type: string, description: '"default" or the targeted version id'}
                file: {type: string, format: binary, description: 'Per-asset limit, checked by `file_size`: 100 MB for an image, 200 MB for a video.'}
      responses:
        '201':
          description: Uploaded and checked
          content:
            application/json:
              schema:
                type: object
                required: [assetId, checks]
                properties:
                  assetId: {type: string}
                  checks: {type: array, items: {$ref: '#/components/schemas/Check'}}
        '400': {$ref: '#/components/responses/ValidationFailed'}
        '401': {$ref: '#/components/responses/Unauthorised'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/ChecksFailed'}
  /v1/campaigns/{campaignId}/submit:
    post:
      tags: [Partner — Campaigns]
      operationId: submitCampaign
      summary: Submit for retailer approval (or auto-approve if the advertiser doesn't require it)
      parameters: [{$ref: '#/components/parameters/CampaignId'}]
      responses:
        '200': {description: Submitted, content: {application/json: {schema: {$ref: '#/components/schemas/CampaignStatus'}}}}
        '401': {$ref: '#/components/responses/Unauthorised'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ChecksFailed'}
  /v1/campaigns/{campaignId}/status:
    get:
      tags: [Partner — Campaigns]
      operationId: getCampaignStatus
      summary: Approval status and rejection reason
      parameters: [{$ref: '#/components/parameters/CampaignId'}]
      responses:
        '200': {description: Status, content: {application/json: {schema: {$ref: '#/components/schemas/CampaignStatus'}}}}
        '401': {$ref: '#/components/responses/Unauthorised'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '404': {$ref: '#/components/responses/NotFound'}
...
components:
  securitySchemes:
    partnerToken:
      type: http
      scheme: bearer
      description: Token issued to a connected DSP or tier-2 partner. Identifies the partner; advertiser is passed per request.
    adminSession:
      type: apiKey
      in: cookie
      name: ph_session
      description: |
        Existing HQ Admin session. Who sees what (spec, Who sees each section):
        admin users get everything, including DSP Integration and editing
        advertiser settings; marketing users get Display Types, Playlist
        Management, Advertisers / Inventory (read-only on the per-advertiser
        settings) and Campaign Status; help desk users get none of it.
        Scopes: admin (DSP Integration, saving advertiser settings), approver
        (approve/reject; HQ Admin role by default, open question 39),
        sections (the four shared sections). In the POC there is no cookie —
        the session is a stand-in whose role comes from the POC_ROLE env var
        (hq_admin | hq_marketing | hq_helpdesk).
  parameters:
...
    CampaignId: {name: campaignId, in: path, required: true, schema: {type: string}}
...
  responses:
    Unauthorised: {description: Missing or invalid credentials, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
...
    NotFound: {description: 'Not found, or not visible to this caller', content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    ValidationFailed: {description: 'error.code = validation_failed; details lists each field', content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    VariableNotPermitted: {description: 'error.code = variable_not_permitted; details names each variable', content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    ChecksFailed: {description: 'error.code = checks_failed; details lists each failed check with its reason', content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
...
    Conflict: {description: 'error.code = conflict (e.g. version changed, already submitted, window sold, bidding not open or closed, live not allowed)', content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
...
    RateLimited: {description: 'error.code = rate_limited. Per partner token (default 50 requests/s, bursts of 100), or too many asset uploads in flight at once. Retry-After gives the seconds to wait.', headers: {Retry-After: {schema: {type: integer}}}, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [validation_failed, variable_not_permitted, checks_failed, not_approved, below_floor, advertiser_blocked, category_blocked, not_on_whitelist, not_invited, targeting_not_supported, conflict, has_dependents, unauthorised, forbidden, not_found, rate_limited, internal_error]
            message: {type: string}
            details:
              type: array
              items:
                type: object
                properties:
                  field: {type: string}
                  variable: {type: string}
                  reason: {type: string}
...
    Condition:
      type: object
      required: [source, variable, op, values]
      properties:
        source: {type: string}
        variable: {type: string}
        op: {type: string, enum: [include, match_exactly, exclude_or, exclude_and, equal, not_equal, greater_than, less_than, greater_than_or_equal, less_than_or_equal]}
        values: {type: array, items: {type: string, maxLength: 200}, maxItems: 100}
    TargetingRules:
      description: AND groups of OR conditions (existing Targeting tab structure).
      type: array
      maxItems: 10
      items: {type: array, minItems: 1, maxItems: 20, items: {$ref: '#/components/schemas/Condition'}}
    CampaignBrief:
      type: object
      description: |
        What the advertiser is booking, in the same shape as the platform's
        Campaign Brief tab (Rob, 20 Sep). Every field is optional: an
        advertiser supplies as much as it has. Stored with the campaign and
        shown to the reviewer; nothing here targets or prices anything.
      properties:
        details: {type: string, maxLength: 2000, description: The campaign in the advertiser's own words.}
        landingPageUrl: {type: string}
        promotedProducts: {type: array, maxItems: 50, items: {type: string}}
        skus: {type: array, maxItems: 100, items: {type: string}}
        targetAudiences: {type: array, maxItems: 50, items: {type: string}}
        objective: {type: string, maxLength: 200}
        touchPoints:
          type: array
          maxItems: 2
          items: {type: string, enum: [Digital Signage, Kiosk]}
          description: Digital Signage only for now; Kiosk is accepted for later.
    PricingType: {type: string, enum: [default, localised, personalised, interactive]}
    CampaignCreate:
      type: object
      required: [advertiserId, name, default]
      description: |
        default is mandatory for every submission (decision, 22 Sep,
        superseding the earlier same-day "baseline optional" decision —
        ticket "Make default creative mandatory; retire localised-only
        booking path"): the earlier fallback-free/part-sold submission
        shape (a campaign with targeted versions but no baseline/default,
        leaving stores none of them match unsold to another advertiser) is
        retired. Every slot now goes to a single advertiser, who supplies a
        mandatory default layer plus optional localised/personalised
        layers as upsells on that one purchase.
      properties:
        advertiserId: {type: string}
        name: {type: string, maxLength: 200}
        displayTypeId: {type: string}
        brief: {$ref: '#/components/schemas/CampaignBrief'}
        default:
          type: object
          required: [pricingType]
          description: The untargeted layer every submission must carry — no targeting variables, priced at the floor rate.
          properties:
            pricingType: {$ref: '#/components/schemas/PricingType'}
        targeted:
          type: array
          maxItems: 20
          description: Optional upsells on top of the mandatory default layer — localised and/or personalised targeted versions.
          items:
            type: object
            required: [id, priority, pricingType, rules]
            properties:
              id: {type: string}
              priority: {type: integer}
              pricingType: {$ref: '#/components/schemas/PricingType'}
              rules: {$ref: '#/components/schemas/TargetingRules'}
    ApprovalStatus: {type: string, enum: [draft, awaiting_approval, approved, rejected]}
    Check:
      type: object
      required: [name, passed]
      properties:
        name: {type: string, enum: [file_type, file_size, bitrate, dimensions, aspect_ratio, duration, default_present, targeting_permitted]}
        passed: {type: boolean}
        detail: {type: string}
        assetId: {type: string, description: '"default" or a targeted version id — the specific asset this check ran against. Unset for a campaign-level check (default_present, targeting_permitted).'}
...
    CampaignStatus:
      type: object
      required: [campaignId, status]
      properties:
        campaignId: {type: string}
        status: {$ref: '#/components/schemas/ApprovalStatus'}
        mode: {type: [string, 'null'], enum: [manual, auto, null]}
        reason: {type: [string, 'null']}
        assetVersion: {type: [string, 'null']}
...